requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.2
django-allauth==0.35.0
django-crispy-forms==1.7.0
gunicorn==19.7.1
idna==2.6
oauthlib==2.0.6
Pillow==5.0.0
psycopg2==2.7.4
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
Unipath==1.1
urllib3==1.22
braintree
